pyproject.toml
[tool.poetry]
name = "py-stocktotum"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sanic = "^24.6.0"
apscheduler = "^3.10.4"
alpha-vantage = "^3.0.0"
jinja2 = "^3.1.4"
aiofiles = "^24.1.0"
types-aiofiles = "^24.1.0.20240626"
pytest = "^8.3.4"
yfinance = "^0.2.54"
scipy-stubs = "^1.15.2.0"
numpy = "^2.2.3"

plyer = "^2.1.0"
pandas-stubs = "^2.2.3.250308"
matplotlib = "^3.10.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
flake8 = "^7.0.0"
ruff = "^0.5.0"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
# Dont automatically fix unused imports here
ignore = ["F401"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = [
    "A",
    "B",
    "C",
    "D",
    "E",
    "F",
    "G",
    "I",
    "N",
    "Q",
    "S",
    "T",
    "W",
    "ANN",
    "ARG",
    "BLE",
    "COM",
    "DJ",
    "DTZ",
    "EM",
    "ERA",
    "EXE",
    "FBT",
    "ICN",
    "INP",
    "ISC",
    "NPY",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "YTT",
]

extend-select = ["I001"]

# F841 = Dont remove unused variables on save, it's annoying when refactoring
unfixable = ["F841"]

# Exclude a variety of commonly ignored directories.
exclude = [".git", ".ruff_cache", ".venv", "venv"]

# I dont really care about long lines, unless they are criminally long
line-length = 110

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.8
target-version = "py38"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
